Paginated based library
overview:
each status on the library page is independently queried at 50 entries at a
time. entries are sorted on the 'all' page so that scrolling brings in entries
in the correct order.

changes:
  - renamed `{{library-status-button}}` to `{{library-dropdown}}`
  - renamed route-level library components
  - updated `ember-one-way-controls` to `v1.0.2` due to an issue they fixed.
  - renamed pagination component to `{{pagination-scrolling}}` and moved most
    logic to a mixin, for later creation of a manual pagination component.

notes:
this still doesn't have full functionality due to `jsonapi-authorization` issues.
  - cannot `create` due to `record` not being fulfilled in policy.
  - cannot `update` due to ember-data serializing relationships and `ja-a`
    calling `relationship#update?` which will always fail for us.
also, the library entries are not yet filterable.

asd