Update docker node base image 5.9 -> 7.1
Currently, our base image is locked to 5.9 However, this doesn't
even seem to be a valid tag for it anymore, according to docker
repository.  Will update this to 7.1, it looks like this doesn't
cause any issues with our dependencies, aside the fact that some
have not updated their metadata properly.