Posters now show on media pages (#1138)
* Posters now show on media pages

It's a very dirty and disgusting fix of this. I'll try to clean this up later by de-bootstrapifying the page.

* Made it more stylish

Somehow this broke scoring media when you haven't scored it previously. I have to fix that.

* Improved display on extra small screens

* Fix layout issues with navbar on media pages

* Fix related media images becoming too large at small screens

Now they won't grow super-large on small screens, making you wonder if it's the primary content or something else